Decompose Redis Configuration Options into a Normalized JSON Object
This has some interesting and unavoidable compromises. The empty
key corresponds to how the endpoints have no keys in the string
form of the configuration options. Sure, fine, OK. But we also have
to put the ID of the replication group in here somewhere. I suppose
this is also a bit of futureproofing; other metadata can go in
there. Clients should feel free to believe that _any_ key starting
with an underscore is reserved.